requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
scipy>=1.9.0
imbalanced-learn>=0.9.0
xgboost>=1.6.0
google-genai>=1.0.0
pydantic>=2.0.0
google-genai
werkzeug
